test: extract process_benchmark deps into a pytest fixture
